restructure of api controllers/routes
